+++ models.go
@@ -16,7 +16,7 @@ type Purchase struct {
 	FarmerID             uuid.UUID `json:"farmer_id"`
 	UserName             string    `json:"user_name"`
 	FarmerName           string    `json:"farmer_name"`
-	FarmerChickenBalance int32     `json:"chicken_balance"`
+	FarmerChickenBalance float64   `json:"chicken_balance"`
 	FarmerCashBalance    int32     `json:"cash_balance"`
 }
 
@@ -30,7 +30,7 @@ type Payment struct {
 	FarmerID             uuid.UUID `json:"farmer_id"`
 	UserName             string    `json:"user_name"`
 	FarmerName           string    `json:"farmer_name"`
-	FarmerChickenBalance int32     `json:"chicken_balance"`
+	FarmerChickenBalance float64   `json:"chicken_balance"`
 	FarmerCashBalance    int32     `json:"cash_balance"`
 }
 
@@ -39,7 +39,7 @@ type Farmer struct {
 	CreatedAt      time.Time `json:"created_at"`
 	UpdatedAt      time.Time `json:"updated_at"`
 	Name           string    `json:"name"`
-	ChickenBalance int32     `json:"chicken_balance"`
+	ChickenBalance float64   `json:"chicken_balance"`
 	CashBalance    int32     `json:"cash_balance"`
 }
 
@@ -98,7 +98,7 @@ func databaseFarmerToFarmer(dbFarmer database.Farmer) Farmer {
 		CreatedAt:      dbFarmer.CreatedAt,
 		UpdatedAt:      dbFarmer.UpdatedAt,
 		Name:           dbFarmer.Name,
-		ChickenBalance: dbFarmer.ChickenBalance.Int32,
+		ChickenBalance: dbFarmer.ChickenBalance.Float64,
 		CashBalance:    dbFarmer.CashBalance.Int32,
 	}
 }
